==> 05_hybrid.ipynb <==
# Library

import os
import re
import random
import itertools
import platform
import multiprocessing
import multiprocessing.dummy
import logging
import inspect
from typing import List
from datetime import datetime

import jieba
import nltk
import faiss
import numpy as np
import pandas as pd
import sacrebleu
import regex
from sacrebleu import corpus_bleu


print('Python version:', platform.python_version())
print('Jieba version', jieba.__version__)
print('NLTK version:', nltk.__version__)
print('Numpy version:', np.__version__)
print('Pandas version:', pd.__version__)
print('Sacrebleu version:', sacrebleu.__version__)
print('Regex version:', regex.__version__)


if hasattr(faiss, 'StandardGpuResources'):
    print('faiss-gpu detected!')
else:
    print('faiss-cpu detected!')


SEED = 42

os.environ['PYTHONHASHSEED']=str(SEED)
random.seed(SEED)
np.random.seed(SEED)


OTHERS_PATTERN: re.Pattern = regex.compile(r'\p{So}')

START_DATETIME = str(datetime.now())
logging.basicConfig(filename=f'log_{START_DATETIME}.log',level=logging.DEBUG)

DIMENSION = 300
K = 5
RES = faiss.StandardGpuResources()

CO = faiss.GpuClonerOptions()
CO.useFloat16 = False
CO.usePrecomputed = False
CO.indicesOptions = faiss.INDICES_CPU


# Vector

 Perform nearest neighbors precomputation with all source tokens, so translation speed would be very fast

def load_fasttext_embedding(filepath):
    word_index = 0
    word2id = {}
    id2word = {}
    embedding = []

    with open(filepath, 'r', encoding='utf-8', newline='\n', errors='ignore') as f:
        for i, line in enumerate(f):
            if i == 0:
                header = line.split()
                assert len(header) == 2
                assert DIMENSION == int(header[1])
                embedding = np.empty(
                    (int(header[0]), int(header[1])),
                    dtype=np.float32
                )
            else:
                word, word_vector = line.rstrip().split(' ', 1)
                word_vector = np.fromstring(word_vector, sep=' ')

                word2id[word] = word_index
                id2word[word_index] = word
                embedding[i-1] = word_vector[None]
                word_index += 1

    return {
        'word2id'   : word2id,
        'id2word'   : id2word,
        'embedding' : embedding
    }


vec_ours_en = load_fasttext_embedding('./data/muse/debug/whw5r77gaz/vectors-en.txt')
vec_ours_zh = load_fasttext_embedding('./data/muse/debug/whw5r77gaz/vectors-zh.txt')

vec_wiki_en = load_fasttext_embedding('./data/wikivec/wiki.en.align.vec')
vec_wiki_zh = load_fasttext_embedding('./data/wikivec/wiki.zh.align.vec')


def create_index(distance, src, tgt):
    if distance == 'IP':
        index = faiss.IndexFlatIP(DIMENSION)
    elif distance == 'L2':
        index = faiss.IndexFlatL2(DIMENSION)
    else:
        index = faiss.IndexFlatIP(DIMENSION)

    index.add(tgt)
    index = faiss.index_cpu_to_gpu(RES, 0, index)

    src2trg_dist, src2trg_id = index.search(src, K)

    return {
        'index'         : index,
        'distance_type' : distance,
        'src2trg_dist'  : src2trg_dist,
        'src2trg_id'    : src2trg_id
    }


index_ours_ip = create_index('IP', vec_ours_zh['embedding'], vec_ours_en['embedding'])
# index_ours_l2 = create_index('L2', vec_ours_zh['embedding'], vec_ours_en['embedding'])

index_wiki_ip = create_index('IP', vec_wiki_zh['embedding'], vec_wiki_en['embedding'])
# index_wiki_l2 = create_index('L2', vec_wiki_zh['embedding'], vec_wiki_en['embedding'])


# Check sanity

def check_sanity(index, vec_trg):
    nearest_dist, nearest_id = index['index'].search(vec_trg['embedding'][1000:1010], K)
    words = [vec_trg['id2word'][i] for i in range(1000, 1010)]
    for i in range(len(words)):
        word = words[i]
        print('Word :', word)

        # word_id = vec_trg['word2id'][word]
        nearest_words = [vec_trg['id2word'][nearest_id[i, j]] for j in range(K)]
        nearest_words_dist = nearest_dist[i]

        print('Nearest words:')
        for i in range(K):
            print(f'{nearest_words[i]} ({nearest_words_dist[i]})')
        print('='*60)


check_sanity(index_ours_ip, vec_ours_en)
# check_sanity(index_ours_l2, vec_ours_en)

check_sanity(index_wiki_ip, vec_wiki_en)
# check_sanity(index_wiki_l2, vec_wiki_en)


# Check average top-1 index distance

dist_wiki = np.empty((len(index_wiki_ip['src2trg_dist'])))
for i in range(len(index_wiki_ip['src2trg_dist'])):
    dist_wiki[i] = index_wiki_ip['src2trg_dist'][i][0]


dist_wiki_sr = pd.Series(dist_wiki)
dist_wiki_sr.describe()


dist_ours = np.empty((len(index_ours_ip['src2trg_dist'])))
for i in range(len(index_ours_ip['src2trg_dist'])):
    dist_ours[i] = index_ours_ip['src2trg_dist'][i][0]


dist_ours_sr = pd.Series(dist_ours)
dist_ours_sr.describe()



# Dataset

val_en = pd.read_csv('./data/csv/dev_en.csv')
val_en = val_en['translation_output']

val_tcn = pd.read_csv('./data/csv/dev_tcn.csv', usecols=['text'])
val_tcn = val_tcn['text']
val_tcn = val_tcn.apply(lambda t:OTHERS_PATTERN.sub(' ', t))

test_tcn = pd.read_csv('./data/csv/test_tcn.csv', usecols=['text'])
test_tcn = test_tcn['text']
test_tcn = test_tcn.apply(lambda t:OTHERS_PATTERN.sub(' ', t))


# Create dict for missing token in Vector from Google Translate (unused)

# from googletrans import Translator
# translator = Translator(service_urls=['translate.google.com'])

# def flatten_2d(old_list):
#     new_list = []
#     for sublist in old_list:
#         for item in sublist:
#             new_list.append(item)
#     return new_list

# def translate_token(token):
#     batch_token = []
#     for i in range(0, len(token), 250):
#         if i+250 >= len(token):
#             curr_batch = token[i:len(token)]
#         else:
#             curr_batch = token[i:i+250]
#         string_batch ='\n'.join(curr_batch)
#         batch_token.append(string_batch)

#     translated_token = []
#     for t in batch_token:
#         tr_object = translator.translate(t, src='zh-tw', dest='en')
#         curr_batch_translated = tr_object.text
#         curr_batch_translated = curr_batch_translated.split('\n')
#         translated_token = translated_token + curr_batch_translated

#     return translated_token

# def translate_missing_tokens(vec_src, val, test):
#     # check parrarel token exist
#     parrarel_tokens_path = './data/csv/parrarel_tokens.csv'
#     if os.path.isfile(parrarel_tokens_path):
#         print('Parrarel token detected!')
#         df_parrarel_tokens = pd.read_csv(parrarel_tokens_path)

#         parrarel_tokens = {}
#         for i in df_parrarel_tokens.index:
#             parrarel_tokens[df_parrarel_tokens.loc[i, 'zh']] = df_parrarel_tokens.loc[i, 'en'].lower()
#         return parrarel_tokens

#     # get all tokens
#     sentences = val.to_list() + test.to_list()
#     print('Total sentences:', len(sentences))
#     tokens = []
#     for s in sentences:
#         result = jieba.tokenize(s, mode='default')
#         token = [r[0] for r in result]
#         token = [t for t in token if t != '' and t != ' ']
#         tokens.append(token)
#     for s in sentences:
#         result = jieba.tokenize(s, mode='default')
#         token = [r[0] for r in result]
#         token = [t for t in token if t != '' and t != ' ']
#         tokens.append(token)
#     tokens = flatten_2d(tokens)
#     print('Total token:', len(tokens))
#     tokens = list(set(tokens))
#     print('Total unique token:', len(tokens))

#     # filter token
#     tokens = [t for t in tokens if re.sub(r'[^\u4e00-\u9fff]', '', t) != '']
#     print('Total CJK unicode token:', len(tokens))

#     # get missing tokens in vector
#     missing_tokens = []
#     for t in tokens:
#         try:
#             vec_src['word2id'][t]
#         except KeyError:
#             missing_tokens.append(t)
#     print('Total missing token:', len(missing_tokens))

#     # translate all missing token
#     print('Translating missing token...')
#     translated_token = translate_token(missing_tokens)

#     # save to file
#     df_parrarel_tokens = pd.DataFrame({
#         'zh': missing_tokens,
#         'en': translated_token
#     })
#     df_parrarel_tokens.to_csv(parrarel_tokens_path, index=False)
#     print('Parrarel token saved at', parrarel_tokens_path)

#     parrarel_tokens = {}
#     for i in df_parrarel_tokens.index:
#         parrarel_tokens[df_parrarel_tokens.loc[i, 'zh']] = df_parrarel_tokens.loc[i, 'en'].lower()

#     return parrarel_tokens


# parrarel_tokens = translate_missing_tokens(vec_ours_zh, val_tcn, test_tcn)


# Translate function

def translate_word_ours(word):
    word_id = vec_ours_zh['word2id'][word]
    english_ids = index_ours_ip['src2trg_id'][word_id]

    tr_word = vec_ours_en['id2word'][english_ids[0]]
    tr_dist = index_ours_ip['src2trg_dist'][word_id][0]

    tr_word = tr_word.lower()
    tr_word = re.sub(r'[^a-z]', '', tr_word)

    return tr_word, tr_dist

def translate_word_wiki(word):
    word_id = vec_wiki_zh['word2id'][word]
    english_ids = index_wiki_ip['src2trg_id'][word_id]

    tr_word = vec_wiki_en['id2word'][english_ids[0]]
    tr_dist = index_wiki_ip['src2trg_dist'][word_id][0]

    tr_word = tr_word.lower()
    tr_word = re.sub(r'[^a-z]', '', tr_word)

    return tr_word, tr_dist

def translate_word(word, *args):
    min_range_ours = args[0]
    min_range_wiki = args[1]
    priority = args[2]

    if word != re.sub(r'[^\u4e00-\u9fff]', '', word):
        return word
    else:
        try:
            tr_word_ours, tr_dist_ours = translate_word_ours(word)
        except KeyError:
            tr_word_ours = None
        except Exception as ex:
            tr_word_ours = None
            logging.exception(f'def translate_word_ours')
            logging.exception(f'Word: {word}')
            logging.exception(ex)


        try:
            tr_word_wiki, tr_dist_wiki = translate_word_wiki(word)
        except KeyError:
            tr_word_wiki = None
        except Exception as ex:
            tr_word_ours = None
            logging.exception(f'def translate_word_wiki')
            logging.exception(f'Word: {word}')
            logging.exception(ex)

        if tr_word_ours is None and tr_word_wiki is None:
            # zh word not in both vector 
            return word
        elif tr_word_ours is not None and tr_word_wiki is not None:
            # zh word in both vector
            if priority == 'wiki':
                if tr_dist_ours < min_range_ours:
                    return tr_word_ours
                elif tr_dist_wiki < min_range_wiki:
                    return tr_word_wiki
                else:
                    return word
            else: # ours
                if tr_dist_wiki < min_range_wiki:
                    return tr_word_wiki
                elif tr_dist_ours < min_range_ours:
                    return tr_word_ours
                else:
                    return word
        elif tr_word_ours is not None:
            # zh word only in ours vector
            if tr_dist_ours < min_range_ours:
                return tr_word_ours
            else:
                return word
        elif tr_word_wiki is not None:
            # zh word only in wiki vector
            if tr_dist_wiki < min_range_wiki:
                return tr_word_wiki
            else:
                return word
        else:
            logging.warning("Reached 'else' on def translate_word")
            return word


def tokenize_tcn(text, mode):
    try:
        result = jieba.tokenize(text, mode=mode)
        token = [r[0] for r in result]
    except Exception as ex:
        logging.exception(f'def tokenize_tcn')
        logging.exception(f'Text: {text}')
        logging.exception(ex)
        token = []
    finally:
        return token

def translate_sentence(sentence, tokenize_mode, *args):
    '''
    tokenize_mode = 'default' or 'search'
    '''
    # tokenize
    sentence = sentence.lower()
    token = tokenize_tcn(sentence, mode=tokenize_mode)
    token = [t for t in token if t != '' and t != ' ']

    # translate
    new_token = []
    for t in token:
        new_token.append(translate_word(t, *args))
    new_sentence = ' '.join(new_token)

    return new_sentence


# Translate & eval validation data

def eval(preds: List[str], refs: List[str]) -> float:
    """BLEU score computation.

    Strips all characters belonging to the unicode category "So".
    Tokenize with standard WMT "13a" tokenizer.
    Compute 4-BLEU.

    Args:
        preds (List[str]): List of translated texts.
        refs (List[str]): List of target reference texts.
    """
    preds = [OTHERS_PATTERN.sub(' ', text) for text in preds]
    refs = [OTHERS_PATTERN.sub(' ', text) for text in refs]
    return corpus_bleu(
        preds, [refs],
        lowercase=True,
        tokenize='13a',
        use_effective_order= False
    ).score


tokenize_mode = ['search', 'default']
min_range_ours = [v/10 for v in range(50, 450, 25)]
min_range_wiki = [v/100 for v in range(20, 75, 5)]
priority = ['wiki', 'ours']

translate_param = list(itertools.product(*[tokenize_mode, min_range_ours, min_range_wiki, priority]))
print('Total translate param permutation:', len(translate_param))

translated_val_tcn_list = []
for i in range(len(translate_param)):
    logging.info(f'Start translating validation data with param #{i}')
    logging.info(translate_param[i])
    translated_val_tcn = []
    for j in val_tcn.index:
        translated_val_tcn.append(
            translate_sentence(val_tcn[j], *translate_param[i])
        )
    translated_val_tcn_list.append(translated_val_tcn)


best_index = 0
best_bleu_score = 0
for i in range(len(translated_val_tcn_list)):
    bleu_score = eval(translated_val_tcn_list[i], val_en.to_list())
    logging.info(f'Bleu score for param {i}: {bleu_score}')
    print(f'Bleu score for param {i}: {bleu_score}')
    if bleu_score > best_bleu_score:
        best_index = i
        best_bleu_score = bleu_score


# Translate test data

print(f'Start translating test data with best param #{best_index}')
print(translate_param[best_index])

logging.info(f'Start translating test data with best param #{best_index}')
logging.info(translate_param[best_index])

translated_test_tcn = []

for i in test_tcn.index:
    translated_test_tcn.append(translate_sentence(test_tcn[i], *translate_param[best_index]))


df_submission = pd.DataFrame(translated_test_tcn, columns=['translation_output'])
df_submission


df_submission.to_csv(f'./submission_{START_DATETIME}.csv', index=False)
